requirements.txt
numpy
opencv-python
Pillow
torch
torchvision
facenet-pytorch
hdbscan
faiss-cpu
scikit-learn
matplotlib
seaborn
umap-learn
tqdm
glob2
python-dotenv
